Compiles. Need map to be Some, only put map in def
Only need to add a map to the ctxt when you are in struct def
and are defining a new struct type. In that case, you put the
new struct's name, its type, and the map of the members. The
struct name isn't a variable name even though it is in the ctxt.
It will never confict with variable names because it is uppercase
and variable names have to start with lower case. Struct names
can't be repeated in the same scope.